requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery & Redis
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Data processing
pandas>=2.0,<3.0
requests>=2.31,<3.0

# Caching
django-cacheops>=7.0,<8.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Utilities
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<22.0
